+++ src/main/java/roomescape/service/member/MemberService.java
@@ -2,10 +2,10 @@ package roomescape.service.member;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import roomescape.domain.dto.ReservationWithRank;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.MemberRepository;
 import roomescape.domain.reservation.ReservationRepository;
-import roomescape.domain.reservation.ReservationWithRank;
 import roomescape.exception.InvalidMemberException;
 import roomescape.service.member.dto.MemberReservationResponse;
 import roomescape.service.member.dto.MemberResponse;
